testing Jepsen client

+++ version.sh
@@ -66,8 +66,13 @@ then
         echo "completed $2 test case with time-limit $3"
     fi
 
-    cp /jepsen.etcdemo/store/latest/linear.svg debug/linear.svg
-    cp /jepsen.etcdemo/store/latest/linear.svg debug/history.txt
+    if test -f "/jepsen.etcdemo/store/latest/linear.svg"; then
+        echo "invalid result detected copying the result to repo ... "
+        cp /jepsen.etcdemo/store/latest/linear.svg debug/linear.svg
+        cp /jepsen.etcdemo/store/latest/linear.svg debug/history.txt
+    fi
+
+
 
 
 else
